implement badass generic type checker for most erlang terms
Maybe this is unnecessary, but a thought experiment I did informed
otherwise. When implementing a "black box" erlang application, in
dealing with external input from various sources, of varying
complexity levels, I think it's important to be able to inform the
client of exactly what piece(s) of their input is(are) incorrect
(instead of simply crashing on bad input with no real helpful
information as to why). This type checker fills that void, though I'm
unsure of whether such a thing already exists, or in what capacity I'm
reinventing any standard erlang features. It has so far been very
quick and easy to create, and is useful, so I'm running with it.